Move writing to `sendFD` onto its own queue
This is needed as the read loop is blocked on messages being parsed on `queue` (in order to not add an extra copy), so writing to `sendFD` can cause deadlocks if it's waiting to write (due to a full pipe). If we change the parse to copy, we can remove this queue.